Move cable-rendering logic to shaft base class
